+++ src/components/EditSections/EditUserInfo/components/index.js
@@ -1,2 +1,4 @@
-/* eslint-disable import/prefer-default-export */
+export { default as ChangeUserTypeModal } from './ChangeUserTypeModal';
+export { default as DeleteProfilePictureModal } from './DeleteProfilePictureModal';
+export { default as ExternalLinkModal } from './ExternalLinkModal';
 export { default as ProfilePicture } from './ProfilePicture';